ADE-stable-1.0/src/ade/training_manager/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
click>=8.1.0
datasets>=2.12.0
gitpython>=3.1.0
numpy>=1.24.0
psutil>=5.9.0
requests>=2.28.0
rouge>=1.0.1
scikit-learn>=1.2.0
torch>=2.0.0
transformers>=4.30.0
pyyaml>=6.0.0
rich>=13.0.0
websockets>=11.0.0
GPUtil>=1.4.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
PyQt6>=6.4.0
PyInstaller>=5.7.0
google-cloud-storage>=2.10.0
google-cloud-compute>=1.14.0
azure-mgmt-resource>=21.1.0
azure-identity>=1.12.0
azure-mgmt-compute>=30.0.0
matplotlib>=3.7.0
slack-sdk>=3.21.0
wandb>=0.15.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.23.0 